README: Slight restructure
This commit rewrites some parts of the README.

It removes the section about package bot, because we all should just
forget that it ever existed. Thanks, Discord.

It also removes a section that compared the library against Discordgo.
The section was written 3 years ago so it was very outdated. Instead, it
is replaced with a "Library Highlights" section, which provides much
more insights on the library itself.